requirement.txt
# 基础框架
flask==2.0.0

# 数据库连接
py2neo==2021.2.4

# API通信与异步支持
aiohttp==3.10.11
websockets==13.1
httpx==0.28.1
openai==1.84.0

# 数据处理与分析
pandas>=1.3.0
numpy>=1.20.0
pyahocorasick==2.1.0
pydantic==2.10.6

# 网页爬取与解析
beautifulsoup4>=4.10.0
selenium>=4.1.0
webdriver-manager==4.0.2

# 工具库
python-dotenv==1.0.1
PyYAML==6.0.2
pillow==10.4.0
tqdm==4.67.1

# 辅助依赖
aiohappyeyeballs>=2.4.4
aiosignal>=1.3.1
annotated-types>=0.7.0
anyio>=4.5.2
async-timeout>=5.0.1
distro>=1.9.0
frozenlist>=1.5.0
h11>=0.16.0
hpack>=4.0.0
httpcore>=1.0.9
interchange>=2021.0.4
jsonpatch>=1.33
jsonpointer>=3.0.0
monotonic>=1.6
multidict>=6.1.0
nest-asyncio>=1.6.0
packaging>=24.2
propcache>=0.2.0
pydantic_core>=2.27.2
Pygments>=2.19.1
spark-ai-python>=0.4.5
tenacity>=9.0.0
yarl>=1.15.2
zstandard>=0.23.0